+++ go/merge/three_way_set_test.go
@@ -68,7 +68,7 @@ func (s *ThreeWaySetMergeSuite) TestThreeWayMerge_HandleNil() {
 }
 
 func (s *ThreeWaySetMergeSuite) TestThreeWayMerge_Refs() {
-	strRef := s.vs.WriteValue(types.NewStruct("Foo", types.StructData{"life": types.Number(42)}))
+	strRef := s.vs.WriteValue(types.NewStruct("Foo", types.StructData{"life": types.Float(42)}))
 
 	m := items{s.vs.WriteValue(s.create(flatA)), s.vs.WriteValue(s.create(flatB))}
 	ma := items{"r1", s.vs.WriteValue(s.create(flatA))}
